\name{grofit.data}
\Rdversion{1.1}
\alias{grofit.data}
\docType{data}
\title{
Example of growth data
}
\description{
Data frame containing several datasets of growth curves. Format: 1. column: a character vector, naming each experiment which is performed for the different concentrations specified in the 3. column; 2. column: a character vector, giving additional information about the experiments; 3. column: a numeric vector, containing the concentration of a compound under which the different experiments are obtained; 4.-68. column: a numeric vectors, giving the growth values corresponding to the time points specified in \code{\link{grofit.time}}.
}
\usage{data(grofit.data)}
\format{
A data frame with 7 observations on the following 68 variables. Format: 1. column: a character vector, naming each experiment which is performed for the different concentrations specified in the 3. column; 2. column: a character vector, giving additional information about the experiments; 3. column: a numeric vector, containing the concentration of a compound under which the different experiments are obtained; 4.-68. column: a numeric vectors, giving the growth values corresponding to the time points specified in \code{\link{grofit.time}}.
}
\seealso{
\code{\link{grofit.time}}
}
\keyword{datasets}
