\name{grofit.time}
\Rdversion{1.1}
\alias{grofit.time}
\docType{data}
\title{
Example of growth data.
}
\description{
Numeric matrix containing a several datasets of time points corresponding to \code{\link{grofit.data}}.
}
\usage{data(grofit.time)}
\format{
A data frame with 7 observations on the following 65 variables. a numeric vectors giving time points corresponding to the growth values in \code{\link{grofit.data}}.
}
\seealso{
\code{\link{grofit.data}}
}
\keyword{datasets}