\name{summary.drFit}
\Rdversion{1.1}
\alias{summary.drFit}
\title{
Generic summary function for drFit object
}
\description{
Generic summary function for \code{drFit} object. Extracts the \code{data.frame} \code{drTable} generated by \code{\link{drFit}}.
}
\usage{
\method{summary}{drFit}(object, ...)
}
\arguments{
  \item{object}{
Object of class \code{drFit}.
}
  \item{\dots}{
Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.
}
}
\value{
\item{data.frame}{1. column: character, naming the experiment; 2. column: logical, indicating wether a logarithmic transformation was applied to concentration data; 3. column, logical, indicating wether a logarithmic transformation was applied to growth parameter; 4. column, numeric, number of bootstrap samples used; 5.-8. column: numeric, estimated EC50 value and respecting reponse value (5.-6.), also in original scale if a transformation is applied (7.-8.); 9.-10. column: numeric, mean and standard dev. from bootstrap sample; 11.-14. column: 90 and 95 percent confidence interval from bootstrap sample; 15. column: mean EC50 of bootstrap sample in original scale; 16.-19. 90 and 95 percent confidence interval in original scale}
}
\seealso{
\code{\link{drFit}}, \code{\link{summary.drFitSpline}}, \code{\link{summary.drBootSpline}}
}
