\name{summary.gcFit}
\Rdversion{1.1}
\alias{summary.gcFit}
\title{
Generic summary function for gcFit object
}
\description{
Generic summary function for \code{gcFit} object.
}
\usage{
\method{summary}{gcFit}(object, ...)
}
\arguments{
  \item{object}{
Object of class \code{gcFit}.
}
  \item{\dots}{
Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.
}
}
\value{
\item{data.frame}{1. column: character, naming the experiment for which a dose-response curve shall be estimated; 2. column: character, additional information about the experiment; 3. column: numeric, concentration of substrate; 4. column: logical, reliability flag, 5. column: character, naming a parametric model, 6. column: logical, indicating wether a logarithmic transformation was applied to the time points of the growth data; 7. column: logical, indicating wether a logarithmic transformation was applied to the growth values of the growth data; 8. column: number of bootstrap samples used; 9.-12. column: estimates of \eqn{\mu}, \eqn{\lambda}, \eqn{A} and \eqn{integral} from growth curve fit; 13.-15. column: respecting standard deviation for \eqn{\mu}, \eqn{\lambda}, \eqn{A}; 16.-21. column: 90 percent confidence intervals for\eqn{\mu}, \eqn{\lambda}, \eqn{A} ; 22.-27. column: 95 percent confidence intervals for \eqn{\mu}, \eqn{\lambda}, \eqn{A}; 28.-31. column: estimates of \eqn{\mu}, \eqn{\lambda}, \eqn{A} and \eqn{integral} from growth curve spline fit; 32.-35. column: mean of \eqn{\mu}, \eqn{\lambda}, \eqn{A} and \eqn{integral} from bootstrap sample; 36.-39. column: respecting standard deviation; 40.-47. column: 90 percent confidence intervals for \eqn{\mu}, \eqn{\lambda}, \eqn{A} and \eqn{integral}; 48.-55. column: 95 percent confidence intervals for \eqn{\mu}, \eqn{\lambda}, \eqn{A} and \eqn{integral}.}
}
\details{
Columns 1.-8. are generated by \code{\link{gcFit}}; columns 9.-27. are generated by \code{\link{summary.gcFitModel}}; columns 28.-31. are generated by \code{\link{summary.gcFitSpline}}; columns 32.-55. are generated by \code{\link{summary.gcBootSpline}}.
}
\seealso{
\code{\link{gcBootSpline}}, \code{\link{gcFit}}, \code{\link{gcFitSpline}}, \code{\link{gcFitModel}}
}
